Refactor the migration module
