+++ src/many-ledger/src/module.rs
@@ -1,5 +1,6 @@
 use crate::json::InitialStateJson;
 use crate::migration::Migration;
+use crate::migration::data::DataMethods;
 use crate::{error, storage::LedgerStorage};
 use coset::{CborSerializable, CoseKey, CoseSign1};
 use many_error::{ManyError, ManyErrorCode};
