+++ src/many-ledger/src/storage.rs
@@ -1,15 +1,14 @@
 use crate::error;
 #[cfg(feature = "migrate_blocks")]
 use crate::migration;
+use crate::migration::data::DataMethods;
 use crate::migration::{run_migrations, Migration};
-use crate::module::{
-    validate_account, ACCOUNT_TOTAL_COUNT_INDEX, NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX,
-};
+use crate::module::validate_account;
 use many_error::ManyError;
 use many_identity::Address;
 use many_modules::abci_backend::AbciCommitInfo;
 use many_modules::account::features::FeatureInfo;
-use many_modules::data::{DataIndex, DataInfo, DataValue};
+use many_modules::data::{DataIndex, DataInfo};
 use many_modules::{account, events, idstore, EmptyReturn};
 use many_protocol::ResponseMessage;
 use many_types::ledger::{Symbol, TokenAmount};
@@ -260,7 +259,7 @@ pub(super) fn key_for_multisig_transaction(token: &[u8]) -> Vec<u8> {
 
 pub struct LedgerStorage {
     symbols: BTreeMap<Symbol, String>,
-    persistent_store: merk::Merk,
+    pub(crate) persistent_store: merk::Merk,
 
     /// When this is true, we do not commit every transactions as they come,
     /// but wait for a `commit` call before committing the batch to the
@@ -603,13 +602,6 @@ impl LedgerStorage {
         }
     }
 
-    pub fn data_attributes(&self) -> Option<BTreeMap<DataIndex, DataValue>> {
-        self.persistent_store
-            .get(DATA_ATTRIBUTES_KEY)
-            .expect("Error while reading the DB")
-            .map(|x| minicbor::decode(&x).unwrap())
-    }
-
     pub fn data_info(&self) -> Option<BTreeMap<DataIndex, DataInfo>> {
         self.persistent_store
             .get(DATA_INFO_KEY)
@@ -749,7 +741,7 @@ impl LedgerStorage {
             ],
         };
 
-        self.count_addresses(from, to, amount.clone(), symbol);
+        self.update_data_attributes(from, to, amount.clone(), symbol);
 
         self.persistent_store.apply(&batch).unwrap();
 
@@ -765,55 +757,6 @@ impl LedgerStorage {
         }
 
         Ok(())
-    }
-
-    fn count_addresses(
-        &mut self,
-        from: &Address,
-        to: &Address,
-        amount: TokenAmount,
-        symbol: &Address,
-    ) {
-        if let Some(mut attributes) = self.data_attributes() {
-            let key_to = key_for_account_balance(to, symbol);
-            if self
-                .persistent_store
-                .get(&key_to)
-                .expect("Error communicating with the DB")
-                .is_none()
-            {
-                attributes
-                    .entry(*ACCOUNT_TOTAL_COUNT_INDEX)
-                    .and_modify(|x| {
-                        if let DataValue::Counter(count) = x {
-                            *count += 1;
-                        }
-                    });
-                attributes
-                    .entry(*NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX)
-                    .and_modify(|x| {
-                        if let DataValue::Counter(count) = x {
-                            *count += 1;
-                        }
-                    });
-            }
-            let balance_from = self.get_balance(from, symbol);
-            if balance_from == amount {
-                attributes
-                    .entry(*NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX)
-                    .and_modify(|x| {
-                        if let DataValue::Counter(count) = x {
-                            *count -= 1;
-                        }
-                    });
-            }
-            self.persistent_store
-                .apply(&[(
-                    DATA_ATTRIBUTES_KEY.to_vec(),
-                    Op::Put(minicbor::to_vec(attributes).unwrap()),
-                )])
-                .unwrap();
-        }
     }
 
     pub fn hash(&self) -> Vec<u8> {
